Major changes to Item-class
The first two changes are probably the biggest ones of the commit.

I removed all the overload constructors, making implementors specify null, 0, or false, for their constructor parameters as it is too tedious and unnecessary to specify all possible combinations of the Item constructor.

Items will now have to use Item#getPlugin to access the plugin from the parent, this is useful as extending items can now override and change the method signature if they need to return a type corresponding to their own plugin.

This next change could affect implementing plugins, but I don't believe the methods used are used outside the library.

ItemUtils' methods were reordered and some method signatures were changed.

These next ones are less important and do not affect currently existing and implementing plugins.

I moved all the NamespacedKeys referencing the ItemStack item-key storage location into ItemUtils, and now through ItemUtils#addItemKey any ItemStack can be made into a "FoxItem" not sure if this was a good change or not as this gives an odd amount of control to implementors, allowing them to make ItemStacks that were never supposed to be tagged with a certain item-key tagged, and allowing them to identify in the system as an instance of that item.

Finally, necessary changes were made to other classes that may have been effected by these changes (ProjectileHitActionListener and ThyTestItem).